requirements.txt
PyQt5==5.15.10
speechrecognition==3.10.0
pyttsx3==2.90
requests==2.31.0
pyaudio==0.2.14
pycaw>=20230407
psutil==5.9.5
Pillow>=10.0.0
pyautogui==0.9.54
elevenlabs>=0.2.26
pygame>=2.5.0
wikipedia>=1.4.0
beautifulsoup4>=4.12.0
win10toast>=0.9
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
spotipy>=2.23.0
cachetools>=5.3.0
python-dateutil>=2.8.2

